requirements.txt
aiohttp
azure-servicebus
azure-identity
